requirements.txt
aiohttp
#ffmpeg-python
uvloop
python-dotenv